Fix few minor issues in -core and -net

+++ reactor-core/src/main/java/reactor/core/Dispatcher.java
@@ -95,7 +95,7 @@ public interface Dispatcher extends Executor, Resource {
 	 * @return the remaining capacity if supported otherwise it returns a negative value.
 	 * @since 2.0
 	 */
-	int backlogSize();
+	long backlogSize();
 
 
 	/**
