+++ reactor-core/src/main/java/reactor/core/dispatch/TraceableDelegatingDispatcher.java
@@ -112,7 +112,7 @@ public class TraceableDelegatingDispatcher implements Dispatcher {
 	}
 
 	@Override
-	public int backlogSize() {
+	public long backlogSize() {
 		return delegate.backlogSize();
 	}
