+++ reactor-net/src/main/java/reactor/io/net/impl/netty/NettyEventLoopDispatcher.java
@@ -85,7 +85,7 @@ public class NettyEventLoopDispatcher extends MultiThreadDispatcher {
 	}
 
 	@Override
-	public int backlogSize() {
-		return Integer.MAX_VALUE;
+	public long backlogSize() {
+		return Long.MAX_VALUE;
 	}
 }
